// csp-angular19R&D/src/app/models/css-batch-model.ts
export class CssBatchModel {
  id?: number;
  frequency?: string;
  sequence?: number;
  year?: number;
  starT_DATE?: Date;
  enD_DATE?: Date;
  status?: string;
  category?: string;
  createD_BY?: string;
  createD_DATE?: Date;
  updateD_BY?: string;
  updateD_DATE?: Date;
  isactive?: boolean;
  totaL_RECORDS?: number;
  pending?: number;
  verified?: number;
  rejected?: number;
  surveY_SENT?: number;
  surveY_RECD?: number;
}